fix: LocationFromKakaoDTO의 json 파일을 제대로 못 읽는 버그 수정

+++ src/main/java/com/ktc/togetherPet/model/dto/kakaoMap/LocationFromKakaoDTO.java
@@ -17,7 +17,7 @@ public record LocationFromKakaoDTO(
     }
 
     public long getAdministrativeCode() {
-        return Integer.parseInt(getAdministrativeDocuments().code());
+        return Long.parseLong(getAdministrativeDocuments().code());
     }
 
     private Documents getAdministrativeDocuments() {
